requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-genai
pydantic